[front] checkout: move payment logic to lib/api/checkout
